+++ PluboRoutes/Endpoint/GetEndpoint.php
@@ -2,13 +2,11 @@
 namespace PluboRoutes\Endpoint;
 
 /**
- * An Endpoint describes a route and its parameters.
+ * An Endpoint with GET method.
  *
  */
-final class GetEndpoint implements EndpointInterface
+final class GetEndpoint extends Endpoint
 {
-    use EndpointTrait;
-
     /**
      * Constructor.
      *
@@ -18,15 +16,6 @@ final class GetEndpoint implements EndpointInterface
      */
     public function __construct(string $namespace, string $path, callable $config, callable $permission_callback = null)
     {
-        $this->namespace = $namespace;
-        $this->path = $path;
-        $this->config = $config;
-        $this->permission_callback = $permission_callback ?? '__return_true';
-        $this->args = array();
-    }
-
-    public function getMethod()
-    {
-        return \WP_REST_Server::READABLE;
+        parent::__construct($namespace, $path, $config, $permission_callback, \WP_REST_Server::READABLE);
     }
 }
